refactor: rename directory with replication logic
