+++ Kysect.GithubUtils/Kysect.GithubUtils.csproj
@@ -27,7 +27,6 @@
 
 	<ItemGroup>
 		<Folder Include="Models\" />
-		<Folder Include="OrganizationReplicator\PathProvider\" />
 		<Folder Include="RepositorySync\PathToRepositoryFormatters\" />
 	</ItemGroup>
