+++ Kysect.GithubUtils/OrganizationReplication/PathProvider/OrganizationReplicatorPathProvider.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.GithubUtils.RepositorySync;
 
-namespace Kysect.GithubUtils.OrganizationReplicator;
+namespace Kysect.GithubUtils.OrganizationReplication;
 
 public class OrganizationReplicatorPathProvider : IOrganizationReplicatorPathProvider
 {
